:label: (types) Migrate user entity in a new dedicated folder
+ improve game entity with an extend
