+++ src/user/models/user.models.ts
@@ -1,0 +1,26 @@
+import { Field, Int, ObjectType } from '@nestjs/graphql';
+
+@ObjectType()
+export class User {
+  @Field(() => String, { description: 'Id of existing user' })
+  userId: string;
+
+  @Field(() => String, { description: 'Name of the player' })
+  username: string;
+
+  @Field(() => String, {
+    description: 'Role of the player during current game',
+  })
+  role: Role;
+}
+
+@ObjectType()
+export class UserInGame extends User {
+  @Field(() => Int, { nullable: true, description: 'Current vote' })
+  vote: number | null;
+}
+
+export enum Role {
+  SCRUMMASTER = 'SCRUMMASTER',
+  DEVELOPER = 'DEVELOPER',
+}
